memory efficient fq: use it everywhere, delete the old version
Summary:

This PR is the cleanup after #50561. High level, we make the new
definition of fake_quant be the definition used by autograd, but keep the old
function around as a thin wrapper to keep the user facing API the same.

In detail:
1. point `fake_quantize_per_tensor_affine`'s implementation to be `fake_quantize_per_tensor_affine_cachemask`
2. delete the `fake_quantize_per_tensor_affine` backward, autograd will automatically use the cachemask backward
3. delete all the `fake_quantize_per_tensor_affine` kernels, since they are no longer used by anything

Test Plan:

```
python test/test_quantization.py TestFakeQuantize
```

performance testing was done in the previous PR.

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: de70258e9950f1e7a401c52b7fa2082390319690
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51159