[JIT] Add support for default argument values to Torchbind
**Summary**
This commit adds support to Torchbind for specifying default values for
arguments of custom class methods. Similar to pybind, default values for
arguments for a method must be provided at registration time as a list
of `torch::arg` objects containing the desired default values.

**Test Plan**
This commit adds a unit test to `test_torchbind.py` that exercises this
feature.

ghstack-source-id: b105aed1643623731318d19f8302562268119484
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51253